🔄 synced local 'skyvern/' with remote 'skyvern/'
The intention of adding this line before was to hide this from the public facing endpoint this. This was not the correct way to do it, as it breaks SDK generation relying on the file download route. Instead, to hide it, we will edit docs.yml on the OSS side

https://buildwithfern.com/learn/docs/api-references/customize-api-reference-layout#hiding-endpoints